+++ KarhooUISDK/Screens/CheckoutScreen/Components/BottomSheets/FlightNumber/KarhooBottomSheetContentWithTextFieldView.swift
@@ -23,7 +23,7 @@ struct KarhooBottomSheetContentWithTextFieldView: View {
             )
             .padding(.bottom, UIConstants.Spacing.standard)
             .padding(.top, UIConstants.Spacing.standard)
-            KarhooMainButton(title: viewModel.buttonText, isEnabled: $viewModel.isTextfieldValid) {
+            KarhooMainButton(title: $viewModel.buttonText, isEnabled: $viewModel.isTextfieldValid) {
                 viewModel.didTapSave(textFieldValue: viewModel.textFieldText)
             }
         }
