+++ KarhooUISDK/Screens/CheckoutScreen/NewCheckout/Workers/KarhooNewCheckoutPaymentWorker.swift
@@ -194,44 +194,4 @@ final class KarhooNewCheckoutPaymentWorker: NewCheckoutPaymentWorker {
             currency: quote.price.currencyCode
         )
     }
-
-    func requestNewPaymentMethod(showRetryAlert: Bool = false) {
-        // TODO: start add payment method flow
-//        cardRegistrationFlow.setBaseView(view)
-//
-//        let currencyCode = quote.price.currencyCode
-//        let amount = quote.price.intHighPrice
-//        let supplierPartnerId = quote.fleet.id
-//
-//        cardRegistrationFlow.start(
-//            cardCurrency: currencyCode,
-//            amount: amount,
-//            supplierPartnerId: supplierPartnerId,
-//            showUpdateCardAlert: showRetryAlert,
-//            callback: { [weak self] result in
-//                guard let cardFlowResult = result.completedValue() else {
-//                    return
-//                }
-//                self?.handleAddCardFlow(result: cardFlowResult)
-//        })
-    }
-
-    func threeDSecureNonceCheck(
-        quote: Quote,
-        passengerDetails: PassengerDetails,
-        resultCompletion: @escaping (OperationResult<ThreeDSecureCheckResult>) -> Void
-    ) {
-        guard let nonce = getPaymentNonceAccordingToAuthState() else {
-            resultCompletion(.completed(value: .failedToInitialisePaymentService))
-            return
-        }
-        threeDSecureProvider?.threeDSecureCheck(
-            nonce: nonce.nonce,
-            currencyCode: quote.price.currencyCode,
-            paymentAmount: NSDecimalNumber(value: quote.price.highPrice),
-            callback: { result in
-                resultCompletion(result)
-            }
-        )
-    }
 }
